server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.1
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
cryptography==44.0.0
Pillow==11.1.0
imagehash==4.3.1
pdqhash==0.2.8
qrcode[pil]==8.0
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
aiofiles==24.1.0
boto3==1.36.2
pyotp==2.9.0
